details with router
