+++ src/App.js
@@ -4,9 +4,7 @@ import Main from './Main';
 class App extends Component {
   render() {
     return (
-      <div className="App">
-        <Main />
-      </div>
+      <Main />
     );
   }
 }
