--- src/BookDetail/style.scss
@@ -1,111 +1,0 @@
-@import './src/variables/colors';
-@import './src/variables/sizes';
-
-.book-img {
-  height: 272px;
-  padding-right: 21px;
-  width: 191px;
-}
-
-.book-no-img {
-  background: url(../assets/default_book.svg) center no-repeat;
-  background-size: 111px 111px;
-  background-color: $alto;
-  content: '';
-  height: $not-found-img-book;
-  padding: 98px 58px;
-  width: $not-found-img-book;
-  z-index: -1;
-  margin-right: 21px;
-}
-
-.detail {
-  @extend %container;
-  padding-top: 143px;
-}
-
-%container {
-  display: flex;
-  max-width: 800px;
-  margin: auto;
-  flex-wrap: wrap;
-  
-  &::after {
-    border-bottom: 2px solid $cerulean;
-    content: '';
-    width: 100%;
-    padding-top: 20px;
-    margin-bottom: 35px;
-  } 
-}
-
-.sugestions { 
-  @extend %container;
-  flex-direction: column;
-}
-
-.title {
-  font: bold $text-big Roboto-Bold;
-  margin: 0;
-}
-
-.subtitle {
-  font: $text-medium-big Roboto-Regular;
-  margin: 0;
-  padding-top: 10px;
-}
-
-.descrip {
-  color: $dusty-gray;
-  font: $text-medium Roboto-Regular;
-  width: 563px;
-  padding: 18px 0 48px 0;
-  margin: 0;
-}
-
-.rent-button {
-  background-color: $cerulean;
-  border: none;
-  color: $white;
-  height: 38px;
-  width: 192px; 
-}
-
-.extra-padding {
-  padding-top: 22px; 
-}
-
-.book-sugestion {
-  //height: 124px;
-  margin-right: 54px; 
- 
-  //width: 87px;
-  background: url(../assets/default_book.svg) center no-repeat;
-  background-size: 124px 87px;
-  background-color: $alto;
-  padding: 124px 87px;
-}
-
-.row {
-  display: flex;
-  max-width: 800px;
-  flex-wrap: wrap;
-  margin-top: 16px;
-  padding-left: 40px; 
-}
-
-.background-logo{
-  position: relative;
-
-  & > :first-child::before {
-    //background: url(../assets/default_book.svg) center no-repeat;
-    //background-size: 111px 111px;
-    background-color: $alto;
-    content: '';
-   // height: $not-found-img-book;
-    //padding: 98px 58px;
-   position: absolute;
-   // width: $not-found-img-book;
-    z-index: -1;
-  }
-}
