--- src/Dashboard/style.scss
@@ -1,65 +1,0 @@
-@import './src/variables/colors';
-@import './src/variables/sizes';
-
-.book-logo {
-  height: $book-logo-height;
-  position: relative;
-  width: $book-logo-width; 
-}
-
-.book-title {
-  font: bold $text-medium Roboto-Bold;
-}
-
-.book-author {
-  font: $text-small Roboto-Regular;
-}
-
-.no-img {
-  background: url(../assets/default_book.svg) center no-repeat;
-  background-color: $alto;
-  content: '';
-  height: $not-found-img-book;
-  padding: 54px 27px;
-  width: $not-found-img-book;
-  z-index: -1;
-}
-
-.grid {
-  display: grid;
-  grid-gap: 65px 70px;
-  grid-template-columns: repeat(4, 128px);
-  justify-content: center;
-  padding: 65px 0;
-  text-align: left;
-}
-
-.filter {
-  background: $wild-sand;
-  border: none;
-  color: $dusty-gray;
-  font: italic $text-medium Roboto-Italic;
-  height: $filter-height;
-  width: $filter-width;
-  padding-left: 12px;
-}
-
-.filter-margin {
-  margin-right: 28px;
-}
-
-.search {
-  background-color: $atlantis;
-  border: none;
-  height: 40px; 
-  position: absolute;
-  width: 40px;
-}
-
-.display-wrap {
-  display: flex;
-  flex-wrap: wrap;
-  margin: auto;
-  max-width: 722px;
-  padding-top: 143px;
-}
