+++ src/components/BookImg/index.js
