+++ src/screens/BookDetail/components/Comment/style.scss
@@ -1,0 +1,8 @@
+@import 'src/variables/colors';
+@import 'src/variables/sizes';
+
+.comment-descrip {
+  color: $dusty-gray;
+  font: $text-medium Roboto-Regular;
+  width: $descrip-width;
+}
