+++ src/screens/BookDetail/index.js
@@ -1,7 +1,10 @@
 import React, { Component } from 'react';
-import books from '../Dashboard/books';
+import books from '../../constants/books';
 import './style.css';
-import BookImg from '../Dashboard/components/BookImg'
+import BookImg from '../../components/BookImg';
+import Comment from './components/Comment';
+import FutureComment from './components/FutureComment';
+import { Link } from 'react-router-dom'; 
 
 class BookDetail extends Component {
   bookInfo(book) {
@@ -10,13 +13,13 @@ class BookDetail extends Component {
         <div>
           <BookImg alt="img" src={book.image_url} classNameImg="book-img" classNameNoImg="book-no-img" />
         </div>
-        <div>
+        <div className="max-width">
           <h2 className="title">{book.title}</h2>
           <h3 className="subtitle">{book.author}</h3>
           <h3 className="subtitle extra-padding">{book.year}</h3>
           <h3 className="subtitle">{book.genre}</h3>
           <p className="descrip">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.</p>
-          <button className="rent-button title">Alquilar</button>          
+          <button className="rent-button">Alquilar</button>          
         </div>
       </div>
     )
@@ -26,7 +29,7 @@ class BookDetail extends Component {
     return (
       <div className="sugestions">
         <div>
-          <h2 className="title">Sugerencias</h2>
+          <h2 className="title-color">Sugerencias</h2>
         </div>
         <div className="row">  
           <BookImg alt="img" src={null} classNameNoImg="book-sugestion" />
@@ -34,7 +37,21 @@ class BookDetail extends Component {
       </div>
     )
   }
-  
+
+  comments() {
+    return (
+      <div className="sugestions no-border">
+        <div>
+          <h2 className="title-color">Comentarios</h2>
+        </div>
+        <div className="row">  
+          <FutureComment />
+          <Comment />
+          <Comment />
+        </div>
+      </div>
+    )
+  }
   
   render() {
     let bookSelected =  books.filter (book => book.id === parseInt(this.props.match.params.id, 10));
@@ -45,8 +62,10 @@ class BookDetail extends Component {
 
     return (
       <div>
+        <Link className="volver" to='/'>&lt; Volver</Link>
         {this.bookInfo(bookSelected[0])} 
         {this.sugestions()}
+        {this.comments()}
       </div>
     )
   };
