+++ src/screens/Dashboard/components/BookList/style.scss
@@ -1,0 +1,32 @@
+@import 'src/variables/colors';
+@import 'src/variables/sizes';
+
+.book-link {
+  color: $black;
+  text-decoration: none;
+}
+
+.book-logo {
+  height: $book-logo-height;
+  position: relative;
+  width: $book-logo-width; 
+}
+
+.no-img {
+  background: url(../../../../assets/default_book.svg) center no-repeat;
+  background-color: $alto;
+  content: '';
+  height: $not-found-img-book;
+  padding: 54px 27px;
+  width: $not-found-img-book;
+  z-index: -1;
+}
+
+.grid {
+  display: grid;
+  grid-gap: 65px 70px;
+  grid-template-columns: repeat(4, 128px);
+  justify-content: center;
+  padding: 65px 0;
+  text-align: left;
+}
