+++ src/variables/_colors.scss
@@ -4,3 +4,4 @@ $cerulean: #00ADEE;
 $dusty-gray: #9B9B9B;
 $wild-sand: #F4F4F4;
 $white: #FFF;
+$black: #000;
