+++ src/variables/_sizes.scss
@@ -3,8 +3,17 @@ $text-medium-big: 16px;
 $text-medium: 14px;
 $text-small: 12px;
 
+$book-img-height: 272px;
+$book-img-width: 191px;
 $book-logo-height: 182px;
 $book-logo-width: 128px;
+$book-sugestion-height: 124px;
+$book-sugestion-width: 87px;
+$button-height: 38px;
+$button-width: 192px;
+$comment-height: 73px;
+$comment-width: 665px;
+$descrip-width: 679px;
 $filter-height: 38px;
 $filter-width: 238px;
 $not-found-img-book: 75px;
